requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alpaca-py==0.7.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
finta==1.3
frozenlist==1.3.3
idna==3.4
msgpack==1.0.4
multidict==6.0.4
numpy==1.24.1
pandas==1.5.2
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sseclient-py==1.7.2
typing_extensions==4.4.0
urllib3==1.26.14
websockets==10.4
yarl==1.8.2